fix: do not set cache-control header for *every graphql request*
Since all graphql requests are made to "/", when the express-router provided by
this package was used, it would add cache-control header for every request. This
fix changes that to add cache-control header only if the path configured in
`expressRouter.path` is requested.

It also make cacheControlHeader configurable. So client can provide any falsy
value as `expressRouter.cacheControlHeader` and this header will not be added to
client response.